Read arguments from command line
I updated script that adds CSV files to the sqlite database to read the
arguments including the name of the CSV file from the command line. For
now the path is still set as “../“ so it reads files from the parent
folder, not the current folder. That we will want to change at some
point to make configurable, probably through a config file.