Added counties resource
This returns a CSV file with back wage information for each county in
the US in the format currently used by our D3 JavaScript visualization.
A better approach would probably be to change the visualization to read
JSON instead. Also, the file really should be cashed to disk so
probably it makes more sense to take it out of the API and instead have
it be created by the script we create to update the sqlite database
with new data. That script can also recreate the static CSV or son file
at the same time, using this query, and then we can take this out of
the API.